feature: Se reorganizaron las carpetas, agregando el proyecto anterior a Maqueta
