+++ Maqueta/Models/cliente.py
@@ -1,0 +1,24 @@
+class Cliente(Usuario):
+    
+    def __init__(self) -> None:
+        super().__init__()
+    
+    def verCarta(self):
+        pass
+    def crearPedido(self):
+        pass
+ 
+    def seleccionarProducto(self):
+        pass
+   
+    def ordenarProducto(self):
+        pass
+   
+    def reservarMesa(self):
+        pass
+   
+    def buscarProducto(self):
+        pass
+    
+    def verPedido(self):
+        pass
