+++ Maqueta/View/css/carta.css
