+++ Maqueta/View/registrar.html
