--- Models/administrador.py
@@ -1,18 +1,0 @@
-
-class Administrador(Usuario):
-    def __init__(self) -> None:
-        super().__init__()
-    
-    def modificarPedido(self):
-        pass
-    def crearCarta(self):
-        pass
-    def agregarProducto(self):
-        pass
-    def modificarProducto(self):
-        pass
-    def modificarReserva(self):
-        pass
-    def listarPedido(self):
-        pass
-
